// src/types/enums.ts
export enum MarketStatus {
  NORMAL = 'NORMAL',
  WARNING = 'WARNING',
  DELISTED = 'DELISTED'
}

export enum EventType {
  STREAM = 'STREAM',
  STREAM_3D = 'STREAM_3D',
  VIDEO = 'VIDEO'
}

export enum ReviewStatus {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED'
}

export enum OrderSide {
  BUY = 'BUY',
  SELL = 'SELL'
}

export enum UserChoice {
  CASH_ONLY = 'CASH_ONLY',
  STOCK_CHOSEN = 'STOCK_CHOSEN'
}